Conflicts:
	modules/matterhorn-composer-ffmpeg/src/main/java/org/opencastproject/composer/impl/AbstractCmdlineEncoderEngine.java
	modules/matterhorn-composer-ffmpeg/src/main/java/org/opencastproject/composer/impl/ComposerServiceImpl.java
	modules/matterhorn-composer-ffmpeg/src/test/resources/encodingprofiles.properties
	modules/matterhorn-composer-ffmpeg/src/test/resources/encodingtestprofiles.properties
	modules/matterhorn-composer-gstreamer/src/main/java/org/opencastproject/composer/gstreamer/GStreamerComposerService.java
	modules/matterhorn-composer-gstreamer/src/main/java/org/opencastproject/composer/gstreamer/endpoint/GStreamerComposerRestService.java
	modules/matterhorn-composer-service-api/src/main/java/org/opencastproject/composer/api/EncodingProfile.java

+++ modules/matterhorn-composer-ffmpeg/src/main/java/org/opencastproject/composer/impl/episode/EpisodeEncoderEngine.java
@@ -22,6 +22,7 @@ import org.opencastproject.composer.impl.AbstractEncoderEngine;
 import org.opencastproject.util.ConfigurationException;
 import org.opencastproject.util.PathSupport;
 import org.opencastproject.util.data.Option;
+
 import org.osgi.service.component.ComponentContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -117,7 +118,7 @@ public final class EpisodeEncoderEngine extends AbstractEncoderEngine {
 
   /**
    * Configures the compression engine wrapper for use with an xmlrpc controller.
-   * 
+   *
    * @param properties
    *          the engine configuration
    */
@@ -193,7 +194,7 @@ public final class EpisodeEncoderEngine extends AbstractEncoderEngine {
    * <p>
    * The hostname parameter is gathered from the configuration file <code>episode.properties</code>.
    * </p>
-   * 
+   *
    * @return the xmlrpc hostname
    */
   public String getXmlrpcHost() {
@@ -205,7 +206,7 @@ public final class EpisodeEncoderEngine extends AbstractEncoderEngine {
    * <p>
    * The port parameter is gathered from the configuration file <code>episode.properties</code>.
    * </p>
-   * 
+   *
    * @return the xmlrpc hostname
    */
   public int getXmlrpcPort() {
@@ -217,7 +218,7 @@ public final class EpisodeEncoderEngine extends AbstractEncoderEngine {
    * <p>
    * The path parameter is gathered from the configuration file <code>episode.properties</code>.
    * </p>
-   * 
+   *
    * @return the xmlrpc path on the server
    */
   public String getXmlrpcPath() {
@@ -226,7 +227,7 @@ public final class EpisodeEncoderEngine extends AbstractEncoderEngine {
 
   /**
    * Returns the number of seconds that are between two calls for a status update.
-   * 
+   *
    * @return the number of seconds for the monitoring interval
    */
   public long getMonitoringFrequency() {
@@ -242,7 +243,7 @@ public final class EpisodeEncoderEngine extends AbstractEncoderEngine {
 
   /**
    * {@inheritDoc}
-   * 
+   *
    * @see org.opencastproject.composer.api.EncoderEngine#mux(java.io.File, java.io.File,
    *      org.opencastproject.composer.api.EncodingProfile, java.util.Map)
    */
@@ -257,7 +258,7 @@ public final class EpisodeEncoderEngine extends AbstractEncoderEngine {
 
   /**
    * {@inheritDoc}
-   * 
+   *
    * @see org.opencastproject.composer.api.EncoderEngine#encode(java.io.File,
    *      org.opencastproject.composer.api.EncodingProfile, java.util.Map)
    */
@@ -269,19 +270,18 @@ public final class EpisodeEncoderEngine extends AbstractEncoderEngine {
 
   /**
    * (non-Javadoc)
-   * 
-   * @see org.opencastproject.composer.api.EncoderEngine#extract(java.io.File,
-   *      org.opencastproject.composer.api.EncodingProfile, java.util.Map, long[])
+   *
+   * @see org.opencastproject.composer.api.EncoderEngine#extract(File, EncodingProfile, Map, double...)
    */
   @Override
-  public List<File> extract(File mediaSource, EncodingProfile format, Map<String, String> properties, long... time)
+  public List<File> extract(File mediaSource, EncodingProfile format, Map<String, String> properties, double... time)
           throws EncoderException {
     throw new UnsupportedOperationException("Operation is not supported by episode encoder engine");
   }
 
   /**
    * {@inheritDoc}
-   * 
+   *
    * @see org.opencastproject.composer.api.EncoderEngine#trim(java.io.File,
    *      org.opencastproject.composer.api.EncodingProfile, long, long, java.util.Map)
    */
@@ -294,7 +294,7 @@ public final class EpisodeEncoderEngine extends AbstractEncoderEngine {
 
   /**
    * {@inheritDoc}
-   * 
+   *
    * @see org.opencastproject.composer.impl.AbstractEncoderEngine#getOutputFile(java.io.File,
    *      org.opencastproject.composer.api.EncodingProfile)
    */
@@ -306,7 +306,7 @@ public final class EpisodeEncoderEngine extends AbstractEncoderEngine {
 
   /**
    * Callback from the engine controllers stating that encoding of the given file has been successful.
-   * 
+   *
    * @param file
    *          the track that was encoded
    * @param profile
@@ -319,7 +319,7 @@ public final class EpisodeEncoderEngine extends AbstractEncoderEngine {
   /**
    * Callback from the engine controllers stating that encoding of the given file has has failed for the specified
    * reason.
-   * 
+   *
    * @param file
    *          the file that was encoded
    * @param profile
@@ -333,7 +333,7 @@ public final class EpisodeEncoderEngine extends AbstractEncoderEngine {
 
   /**
    * Callback from the engine controllers stating that encoding of the given file has progressed to the given value.
-   * 
+   *
    * @param file
    *          the file that is being encoded
    * @param profile
@@ -354,7 +354,7 @@ public final class EpisodeEncoderEngine extends AbstractEncoderEngine {
 
   /*
    * (non-Javadoc)
-   * 
+   *
    * @see org.opencastproject.composer.api.EncoderEngine#extract(java.io.File,
    * org.opencastproject.composer.api.EncodingProfile, java.util.Map, long[])
    */
